bump java to 17, bump dependencies, reformat code
