+++ src/main/java/pl/put/CinemaManagement/CinemaManagementApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CinemaManagementApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CinemaManagementApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CinemaManagementApplication.class, args);
+    }
 
 }
