--- src/main/java/pl/put/CinemaManagement/config/CORSConfig.java
@@ -1,25 +1,0 @@
-package pl.put.CinemaManagement.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
-import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurer;
-
-@Slf4j
-@Configuration
-public class CORSConfig implements RepositoryRestConfigurer {
-
-    @Value("${angular.client}")
-    String angularClient;
-
-    @Override
-    public void configureRepositoryRestConfiguration(RepositoryRestConfiguration config) {
-        config.getCorsRegistry()
-                .addMapping("/**")
-                .allowedMethods("GET", "PUT", "POST", "DELETE")
-                .allowedHeaders("*")
-                .allowedOrigins(angularClient);
-        log.info("CORS allowed for: " + angularClient);
-    }
-}
