+++ src/main/java/pl/put/CinemaManagement/file/service/FileService.java
@@ -5,5 +5,6 @@ import java.io.InputStream;
 
 public interface FileService {
     byte[] get(String key) throws IOException;
+
     void put(String key, InputStream fileInputStream) throws IOException;
 }
