+++ src/main/java/pl/put/CinemaManagement/file/service/S3FileService.java
@@ -2,7 +2,6 @@ package pl.put.CinemaManagement.file.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.GetObjectRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.io.FileUtils;
