+++ src/main/java/pl/put/CinemaManagement/model/BasePrice.java
@@ -10,7 +10,8 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
-@Getter @Setter
+@Getter
+@Setter
 @NoArgsConstructor
 @Entity
 @Table(
