+++ src/main/java/pl/put/CinemaManagement/model/FilmShow.java
@@ -8,7 +8,6 @@ import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 import java.sql.Date;
-import java.sql.Timestamp;
 import java.util.List;
 
 @Getter
