+++ src/main/java/pl/put/CinemaManagement/repository/CinemaRepository.java
@@ -3,8 +3,6 @@ package pl.put.CinemaManagement.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.put.CinemaManagement.model.Cinema;
 
-import java.util.List;
-
 public interface CinemaRepository extends CrudRepository<Cinema, Long> {
 
     Cinema findCinemaByName(String name);
