+++ src/main/java/pl/put/CinemaManagement/repository/PromoOfferRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface PromoOfferRepository extends CrudRepository<PromoOffer, Long> {
 
-    PromoOffer findPromoOfferByName (String name);
+    PromoOffer findPromoOfferByName(String name);
 
     @Query("from PromoOffer p where p.id = any " +
             "(select s.promoOffer.id from ClientSegment s where " +
