+++ src/main/resources/application.properties
@@ -26,3 +26,6 @@ keycloak.ssl-required = external
 keycloak.resource=cinema-management-backend
 keycloak.use-resource-role-mappings=true
 keycloak.bearer-only=true
+
+#workaround for springfox to work with spring boot >2.6.0
+spring.mvc.pathmatch.matching-strategy=ant_path_matcher
