--- src/test/java/pl/put/CinemaManagement/CinemaManagementApplicationTests.java
@@ -1,13 +1,0 @@
-package pl.put.CinemaManagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CinemaManagementApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
